image height adjust

+++ lib/base_widget/nfpost_player.dart
@@ -8,7 +8,7 @@ class Player extends StatefulWidget {
   final String url;
   final int id;
   final int index;
-  final void Function(int height) maxHeightCallBack;
+  final void Function(double height) maxHeightCallBack;
 
   const Player({Key key, this.url, this.id, this.index, this.maxHeightCallBack}) : super(key: key);
 
@@ -40,7 +40,7 @@ class _PlayerState extends State<Player> {
     _muteSubject.add(_controller.value.volume == 0.0);
     _controller.initialize().then((_) {
       _controller.setLooping(true);
-      widget.maxHeightCallBack(_controller.value.size.height.toInt());
+      widget.maxHeightCallBack(_controller.value.size.height);
       _controller.seekTo(Duration(milliseconds: 50));
       setState(() {});
     });
